Renames file according to review
